Finsished shiptool, works now wohoo :D

+++ src/main/java/de/prisemut/mbb/shiptool/Schematic.java
@@ -2,13 +2,13 @@ package de.prisemut.mbb.shiptool;
 
 public class Schematic {
 
-    private byte[] blocks;
+    private short[] blocks;
     private byte[] data;
     private short width;
     private short lenght;
     private short height;
 
-    public Schematic(byte[] blocks, byte[] data, short width, short lenght, short height)
+    public Schematic(short[] blocks, byte[] data, short width, short lenght, short height)
     {
         this.blocks = blocks;
         this.data = data;
@@ -20,7 +20,7 @@ public class Schematic {
     /**
      * @return the blocks
      */
-    public byte[] getBlocks()
+    public short[] getBlocks()
     {
         return blocks;
     }
